+++ frontend/src/App.jsx
@@ -1,10 +1,11 @@
 import React, {useContext} from 'react';
-import {UserContext} from './context/UserContext';
 import {BrowserRouter, Routes, Route, Navigate} from 'react-router-dom';
+
+import {UserContext} from './context/UserContext';
 
 import Index from './page/index/Index';
 import ServiceMain from './page/service/main/ServiceMain';
-import OwnerMain from './page/owner/main/OwnerMain';
+import OwnerIndex from './page/owner/index/OwnerIndex';
 
 import './App.css';
 
@@ -14,7 +15,7 @@ export default function App() {
             <Routes>
                 <Route exact="exact" path="/" element={<Index />}/>
                 <Route path="/events/:eventId/*" element={<ServiceMain />}/>
-                <Route path="/owner/*" element={<OwnerMain />}/>
+                <Route path="/owner/*" element={<OwnerIndex />}/>
                 <Route path="/logout" element={<Logout />}/>
             </Routes>
         </BrowserRouter>
