+++ frontend/src/function/BrowserNotification.jsx
@@ -3,18 +3,7 @@ export const sendNotification = (message) => {
         return;
     }
 
-    if (Notification.permission !== "granted") {
-        Notification
-            .requestPermission()
-            .then((permission) => {
-                if (permission === "granted") {
-                    new Notification(message);
-                } else {
-                    alert("모든 알림이 차단되었습니다.");
-                    return;
-                }
-            });
-    } else {
+    if (Notification.permission === "granted") {
         new Notification(message);
     }
 };
